+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/profiling/trace/ProfileThreadSnapshotRecord.java
@@ -31,7 +31,6 @@ import org.apache.skywalking.oap.server.core.storage.annotation.QueryUnifiedInde
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Storage;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
-import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 import static org.apache.skywalking.oap.server.core.source.DefaultScopeDefine.PROFILE_TASK_SEGMENT_SNAPSHOT;
@@ -95,11 +94,7 @@ public class ProfileThreadSnapshotRecord extends Record {
             converter.accept(DUMP_TIME, storageData.getDumpTime());
             converter.accept(SEQUENCE, storageData.getSequence());
             converter.accept(TIME_BUCKET, storageData.getTimeBucket());
-            if (CollectionUtils.isEmpty(storageData.getStackBinary())) {
-                converter.accept(STACK_BINARY, Const.EMPTY_STRING);
-            } else {
-                converter.accept(STACK_BINARY, new String(Base64.getEncoder().encode(storageData.getStackBinary())));
-            }
+            converter.accept(STACK_BINARY, new String(Base64.getEncoder().encode(storageData.getStackBinary())));
         }
     }
 }