+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/type/Convert2Entity.java
@@ -18,9 +18,20 @@
 
 package org.apache.skywalking.oap.server.core.storage.type;
 
+import java.util.function.Function;
+
 /**
  * A function supplier to convert raw data from database to object defined in OAP
  */
 public interface Convert2Entity {
     Object get(String fieldName);
+
+    /**
+     * Use the given type decoder to decode value of given field name.
+     *
+     * @param fieldName   to read value
+     * @param typeDecoder to decode the value
+     * @return decoded value
+     */
+    <T, R> R getWith(String fieldName, Function<T, R> typeDecoder);
 }